requirements.txt
qoro-divi==0.10.0
networkx
matplotlib
numpy
pennylane
dimod
